+++ app/ai/cnn_model.py
