+++ app/database.py
@@ -1,0 +1,20 @@
+from sqlalchemy.ext.declarative import declarative_base
+from sqlalchemy.ext.asyncio import create_async_engine, AsyncSession
+from sqlalchemy.ext.asyncio import async_sessionmaker
+from app.config import DATABASE_URL
+
+# SQLAlchemy 초기화
+async_engine = create_async_engine(DATABASE_URL, echo=True)
+async_session = async_sessionmaker(async_engine, expire_on_commit=False, class_=AsyncSession)
+Base = declarative_base()
+
+# DB 의존성 주입 헬퍼
+async def get_db():
+    async with async_session() as session:
+        yield session
+
+# 데이터베이스 초기화 함수
+async def init_database():
+    from models import Base
+    async with async_engine.begin() as conn:
+        await conn.run_sync(Base.metadata.create_all)
